requirements.txt
Django==2.1.7
django-bootstrap3==11.0.0
django-debug-toolbar==1.11
django-extensions==2.1.6
gunicorn==19.9.0
pytz==2018.9
six==1.12.0
sqlparse==0.3.0
